requirements.txt
google-cloud-tasks
audioread==3.0.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
Flask==2.3.2
fonttools==4.38.0
future==0.18.3
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0
llvmlite==0.39.1
MarkupSafe==2.1.2
msgpack==1.0.4
numba==0.56.4
numpy==1.23.5
packaging==23.0
pedalboard==0.7.0
Pillow==9.4.0
platformdirs==3.0.0
pooch==1.7.0
pycparser==2.21
pyloudnorm==0.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.3
threadpoolctl==3.1.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.3.6
zipp==3.15.0
